+++ src/containerapp/azext_containerapp/_params.py
@@ -415,6 +415,16 @@ def load_arguments(self, _):
         c.argument('min_nodes', help="The minimum node count for the workload profile")
         c.argument('max_nodes', help="The maximum node count for the workload profile")
 
+     with self.argument_context('containerapp env workload-profile add') as c:
+        c.argument('workload_profile_type', help="The type of workload profile to add to this environment. Run 'az containerapp env workload-profile list-supported -l <region>' to check the options for your region.")
+        c.argument('min_nodes', help="The minimum node count for the workload profile")
+        c.argument('max_nodes', help="The maximum node count for the workload profile")
+
+    with self.argument_context('containerapp env workload-profile update') as c:
+        c.argument('workload_profile_type', help="The type of workload profile to update. Run 'az containerapp env workload-profile list-supported -l <region>' to check the options for your region.")
+        c.argument('min_nodes', help="The minimum node count for the workload profile")
+        c.argument('max_nodes', help="The maximum node count for the workload profile")
+        
     # Patch
     with self.argument_context('containerapp patch list') as c:
         c.argument('resource_group_name', arg_type=resource_group_name_type)